fix: native codes that are not running
