+++ src/index.ts
@@ -1,3 +1,49 @@
+import {NativeModules, Platform} from 'react-native';
+
+const {RNKakaoLogins} = NativeModules;
+
+export type KakaoOAuthToken = {
+  accessToken: string;
+  refreshToken: string;
+  idToken: string;
+  accessTokenExpiresAt: Date;
+  refreshTokenExpiresAt: Date;
+  scopes: string[];
+};
+
+export type KakaoAccessTokenInfo = {
+  accessToken: string;
+  expiresIn: string;
+};
+
+export type KakaoProfile = {
+  id: string;
+  email: string;
+  nickname: string;
+  profileImageUrl: string;
+  thumbnailImageUrl: string;
+  phoneNumber: string;
+  ageRange: string;
+  birthday: string;
+  birthdayType: string;
+  birthyear: string;
+  gender: string;
+  isEmailValid: boolean;
+  isEmailVerified: boolean;
+  isKorean: boolean;
+  ageRangeNeedsAgreement?: boolean;
+  birthdayNeedsAgreement?: boolean;
+  birthyearNeedsAgreement?: boolean;
+  emailNeedsAgreement?: boolean;
+  genderNeedsAgreement?: boolean;
+  isKoreanNeedsAgreement?: boolean;
+  phoneNumberNeedsAgreement?: boolean;
+  profileNeedsAgreement?: boolean;
+};
+
+export type KakaoProfileNoneAgreement = {
+  id: string;
+};
 export type KakaoOAuthWebToken = {
   access_token: string;
   expires_in: number;
@@ -23,7 +69,7 @@ export type KakaoProfileWebType = {
 
 export const login = async (
   props?: KaKaoLoginWebType,
-): Promise<KakaoOAuthWebToken> => {
+): Promise<KakaoOAuthWebToken | KakaoOAuthToken> => {
   if (!props) {
     throw new Error('Web parameters are not provided');
   }
@@ -46,14 +92,33 @@ export const login = async (
     .join('&');
 
   try {
-    const result = await fetch('https://kauth.kakao.com/oauth/token', {
-      method: 'post',
-      body: queryString,
-      headers: {
-        'Content-type': 'application/x-www-form-urlencoded;charset=utf-8',
-      },
-    });
-    return result.json();
+    if (Platform.OS === 'web') {
+      const result = await fetch('https://kauth.kakao.com/oauth/token', {
+        method: 'post',
+        body: queryString,
+        headers: {
+          'Content-type': 'application/x-www-form-urlencoded;charset=utf-8',
+        },
+      });
+
+      return result.json();
+    }
+
+    return RNKakaoLogins.login();
+  } catch (err) {
+    throw err;
+  }
+};
+
+export const loginWithKakaoAccount = async (): Promise<KakaoOAuthToken> => {
+  if (Platform.OS === 'web') {
+    throw new Error('Web does not support `loginWithKakaoAccount`');
+  }
+
+  try {
+    const result: KakaoOAuthToken = await RNKakaoLogins.loginWithKakaoAccount();
+
+    return result;
   } catch (err) {
     throw err;
   }
@@ -61,13 +126,18 @@ export const login = async (
 
 export const logout = async (tokenWeb?: string): Promise<string> => {
   try {
-    const result = await fetch('https://kapi.kakao.com/v1/user/logout', {
-      method: 'post',
-      headers: {
-        Authorization: `Bearer ${tokenWeb}`,
-      },
-    });
-    return result.json();
+    if (Platform.OS === 'web') {
+      const result = await fetch('https://kapi.kakao.com/v1/user/logout', {
+        method: 'post',
+        headers: {
+          Authorization: `Bearer ${tokenWeb}`,
+        },
+      });
+
+      return result.json();
+    }
+
+    return RNKakaoLogins.logout();
   } catch (err) {
     throw err;
   }
@@ -75,13 +145,18 @@ export const logout = async (tokenWeb?: string): Promise<string> => {
 
 export const unlink = async (tokenWeb?: string): Promise<string> => {
   try {
-    const result = await fetch('https://kapi.kakao.com/v1/user/unlink', {
-      method: 'post',
-      headers: {
-        Authorization: `Bearer ${tokenWeb}`,
-      },
-    });
-    return result.json();
+    if (Platform.OS === 'web') {
+      const result = await fetch('https://kapi.kakao.com/v1/user/unlink', {
+        method: 'post',
+        headers: {
+          Authorization: `Bearer ${tokenWeb}`,
+        },
+      });
+
+      return result.json();
+    }
+
+    return RNKakaoLogins.unlink();
   } catch (err) {
     throw err;
   }
@@ -91,14 +166,33 @@ export const getProfile = async (
   token?: string,
 ): Promise<KakaoProfileWebType> => {
   try {
-    const result = await fetch('https://kapi.kakao.com/v2/user/me', {
-      method: 'get',
-      headers: {
-        Authorization: `Bearer ${token}`,
-        'Content-type': 'application/x-www-form-urlencoded;charset=utf-8',
-      },
-    });
-    return result.json();
+    if (Platform.OS === 'web') {
+      const result = await fetch('https://kapi.kakao.com/v2/user/me', {
+        method: 'get',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-type': 'application/x-www-form-urlencoded;charset=utf-8',
+        },
+      });
+
+      return result.json();
+    }
+
+    return RNKakaoLogins.getProfile();
+  } catch (err) {
+    throw err;
+  }
+};
+
+export const getAccessToken = async (): Promise<KakaoAccessTokenInfo> => {
+  if (Platform.OS === 'web') {
+    throw new Error('Web does not support `getAccessToken`');
+  }
+
+  try {
+    const result: KakaoAccessTokenInfo = await RNKakaoLogins.getAccessToken();
+
+    return result;
   } catch (err) {
     throw err;
   }
